Add more 'class' to the ElconCharger class
We move to object method calls rather than function calls throughout.

As well as the message's decoded values for voltage, current and so on, we
also store the `pf`, `r`, `dp` and `priority` values as object properties.
These don't seem to change, at least in use so far, and they're probably
better left as they are.

We're also standardising on four space indents, in accordance with PEP8.

Signed-off-by: Paul Wayper <paulway@mabula.net>