Add a module for simulating a battery and a charger.
This module provides the `Battery` and `Charger` classes, that
simulate the operation of a battery and charger respectively.

The `Battery` is instantiated with a capacity (in amp-hours, as a
float) and number of cells (as an integer).  It defaults to 50%
charged.  It can be charged and discharged by applying amps for a
number of seconds, and charging will also take the voltage that
the battery is being charged at.  The voltage of the battery
corresponds to its state of charge, using a cubic function that
(roughly) simulates the charge/discharge curve of a lithium-ion
battery.  It cannot be over-charged or over-discharged (limiting
to fully charged or fully discharged).

The `Charger` is instantiated with an instance of a battery given
by the class above.  It can then be set to a given voltage and
current, and told to charge the attached battery for a given number
of seconds.  Charging is finished when the battery's voltage goes
above the charger's set voltage - there is no 'constant voltage'
phase of the discharge in this simulation.

Signed-off-by: Paul Wayper <paulway@mabula.net>