Allow ElconUtils to work for different users
The ElconUtils object needed to work for both the manager and
the charger.  For the manager the source ID needs to be the
manager, and so forth; so the ElconUtils object needed to know
which source and destination it was encoding and decoding for.
Instead of having to pass these as arguments each time, we set
these at initialisation.

Signed-off-by: Paul Wayper <paulway@mabula.net>