Very basic simulation of an Elcon charger
This module contains two classes.

The simpler one is `ChargerDriver`.  When given a `can` module `Bus`
object, this will emit a message every second from the manager to the
charger to tell the charger to emit a given voltage and current.  The
driver's voltage, amperage, maximum wattage draw, efficiency, and update
time can be set.  The driver must be started using `driver.start()`
before it will emit messages.

The more complex one is `ElconCharger`.  This is given a `Bus` object
and a `Charger` object from the `battery` module, which in turn must be
given a `Battery` object.  The ElconCharger then listens on the CAN
bus for messages from the manager to the charger.  When it gets a
correct message, it turns on and charges the battery.  If a message is
not received within `update_timeout` seconds, the charger deactivates
and does not charge the battery.

All this is accomplished using the `asyncio` module.

This is somewhat of a work in progress, changes may happen in the
future.

Signed-off-by: Paul Wayper <paulway@mabula.net>