Change ElconUtils to 'our' bus ID
A CANBUS device usually has a fixed ID, which it receives messages for
and sends messages from.  The destination of outgoing messages can
vary, though, and is not always fixed.  So the idea of having a single
'source' and 'destination' address for the ElconUtils class didn't
really make a lot of sense.

I've changed this over to having an 'our_id', which is the single byte
identifier for this device.  Messages to other IDs are ignored, and this
is used as the source for ourgoing messages.  The major change in the
code is that the `pack_command` method now takes a `pkt_dest` byte that
is the destination ID for the packet.

Signed-off-by: Paul Wayper <paulway@mabula.net>