Moving the Charger ops to the ElconCharger class
The `Charger` class was a bit redundant, as the ElconCharger
class did most of its work.  So I've merged that into the
ElconCharger class entirely.

This also contains the start of the work for the ElconCharger
to send CANBUS messages of its status, and the ChargerDriver
to accept and print them.  This includes supporting the
`timeout` flag, and cutting the charger's output amps to zero
when the battery is fully charged (i.e. its voltage is greater
than the charger's voltage).

Signed-off-by: Paul Wayper <paulway@mabula.net>